Move `useEffect` for `ratings` to `Ratings FC`
Addresses #49 (Breakout `useEffect` calls into individual components)
Project: 'Breakdown `Cockpit` into subcomponents'
