+++ client/src/components/Ratings/Ratings.js
@@ -1,8 +1,10 @@
 import { Typography } from '@material-ui/core';
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import CreateRating from './CreateRating/CreateRating';
 
 import Rating from './Rating/Rating';
+import envs from '../../envs/envs';
+import axios from 'axios';
 
 import Grid from '@material-ui/core/Grid';
 
@@ -23,6 +25,27 @@ const useStyles = makeStyles((theme) => ({
 
 export default function Ratings(props) {
 	const classes = useStyles();
+
+	const [ratings, setRatings] = useState([]);
+
+	useEffect(() => {
+		getRatings();
+		ratingCreated();
+	}, []);
+
+	async function getRatings() {
+		const ratingsRes = await axios.get(`${envs}/ratings`);
+		setRatings(ratingsRes.data);
+	}
+
+	async function ratingCreated() {
+		getRatings();
+	}
+
+	async function ratingDeleted() {
+		getRatings();
+	}
+
 	return (
 		<div>
 			<Typography variant='h4'>Ratings</Typography>
@@ -32,8 +55,8 @@ export default function Ratings(props) {
 				className={clsx('grid-container', classes.root)}
 				style={{ marginBottom: '1rem' }}
 			>
-				{props.ratings &&
-					props.ratings.map((rating, index) => (
+				{ratings &&
+					ratings.map((rating, index) => (
 						<Grid
 							item
 							xs={12}
@@ -48,13 +71,13 @@ export default function Ratings(props) {
 								rating={rating.rating}
 								key={index}
 								id={rating._id}
-								deleteRating={props.cockpitDeleteCallback}
+								deleteRating={ratingDeleted}
 							/>
 						</Grid>
 					))}
 			</Grid>
 
-			<CreateRating ratingCreated={props.cockpitCreateCallback} />
+			<CreateRating ratingCreated={ratingCreated} />
 		</div>
 	);
 }
